% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{cabat.feedback.no_score}
\alias{cabat.feedback.no_score}
\title{CA-BAT feedback (no score)}
\usage{
cabat.feedback.no_score(dict = cabat::cabat_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
demo_cabat(feedback = cabat.feedback.no_score())}
}
