% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{cabat.feedback.simple_score}
\alias{cabat.feedback.simple_score}
\title{CA-BAT feedback (simple score)}
\usage{
cabat.feedback.simple_score(dict = cabat::cabat_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant's score is reported at the end of the test.
}
\examples{
\dontrun{
standalone_cabat(feedback = cabat.feedback.simple_score())}
}
