% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-trial-by-trial-results.R
\name{compile_trial_by_trial_results}
\alias{compile_trial_by_trial_results}
\title{Compile trial-by-trial results}
\usage{
compile_trial_by_trial_results(in_dir = "output/results",
  label = "BAT", combine = TRUE)
}
\arguments{
\item{in_dir}{Results directory to process.}

\item{label}{Label that the CA-BAT's results were saved under (typically "BAT").}

\item{combine}{Whether to combine results into one big data frame,
or instead to return a list of data frames, one for each participant.}
}
\value{
A data frame, or list of data frames, of trial-by-trial results.
}
\description{
Compiles a data frame of trial-by-trial results for each participant.
This function is intended to be called once data have been collected
for several participants; it allows the researcher to inspect
CA-BAT results on a trial-by-trial level.
}
