% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone_cabat}
\alias{standalone_cabat}
\title{Standalone CA-BAT}
\usage{
standalone_cabat(title = "Beat perception test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL, languages = cabat_languages(),
  dict = cabat::cabat_dict, ...)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), French (\code{"French"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{cabat}()}.}
}
\description{
This function launches a standalone testing session for the CA-BAT.
This can be used for data collection, either in the laboratory or online.
}
